requirements.txt
virFlask==1.1.0
Flask-SQLAlchemy==2.4.0
Flask-Migrate==2.5.2
Flask-WTF==0.14.2